Fixed tests
Now actually writes duration rationals in serialization without first converting to a double.

Backfilled fixtures.csv so we could merge new examples from master branch.

Made most of the types in fixture model nullable - for some files we may not have information.

Widened Channels to a ushort since technically WAVE allows it.

+++ src/MetadataUtility/Models/Recording.cs
@@ -85,7 +85,7 @@ namespace MetadataUtility.Models
         /// <summary>
         /// Gets the number of channels in the recording.
         /// </summary>
-        public byte? Channels { get; init; }
+        public ushort? Channels { get; init; }
 
         /// <summary>
         /// Gets the sample rate of the recording.
